Show oldest tickets first
In theory, I should show the most recently updated tickets first, but I
didn't want to use Laravel's `updated_at` because it doesn't keep track
of the contents as they change, and I think that's kind of important for
a customer ticket system.

Given it's a relatively simple project, I didn't really bother implementing
that.